+++ source/extensions/transport_sockets/http_11_proxy/connect.h
@@ -5,6 +5,7 @@
 
 #include "source/common/buffer/buffer_impl.h"
 #include "source/common/common/logger.h"
+#include "source/common/http/http1/balsa_parser.h"
 #include "source/extensions/transport_sockets/common/passthrough.h"
 
 namespace Envoy {
@@ -48,6 +49,45 @@ public:
                         std::shared_ptr<const Upstream::HostDescription> host) const override;
   void hashKey(std::vector<uint8_t>& key,
                Network::TransportSocketOptionsConstSharedPtr options) const override;
+};
+
+// This is a utility class for isValidConnectResponse. It is only exposed for
+// coverage testing purposes. See isValidConnectResponse for intended use.
+class SelfContainedParser : public Http::Http1::ParserCallbacks {
+public:
+  SelfContainedParser() : parser_(Http::Http1::MessageType::Response, this, 2000) {}
+  Http::Http1::CallbackResult onMessageBegin() override {
+    return Http::Http1::CallbackResult::Success;
+  }
+  Http::Http1::CallbackResult onUrl(const char*, size_t) override {
+    return Http::Http1::CallbackResult::Success;
+  }
+  Http::Http1::CallbackResult onStatus(const char*, size_t) override {
+    return Http::Http1::CallbackResult::Success;
+  }
+  Http::Http1::CallbackResult onHeaderField(const char*, size_t) override {
+    return Http::Http1::CallbackResult::Success;
+  }
+  Http::Http1::CallbackResult onHeaderValue(const char*, size_t) override {
+    return Http::Http1::CallbackResult::Success;
+  }
+  Http::Http1::CallbackResult onHeadersComplete() override {
+    headers_complete_ = true;
+    parser_.pause();
+    return Http::Http1::CallbackResult::Success;
+  }
+  void bufferBody(const char*, size_t) override {}
+  Http::Http1::CallbackResult onMessageComplete() override {
+    return Http::Http1::CallbackResult::Success;
+  }
+  void onChunkHeader(bool) override {}
+
+  bool headersComplete() const { return headers_complete_; }
+  Http::Http1::BalsaParser& parser() { return parser_; }
+
+private:
+  bool headers_complete_ = false;
+  Http::Http1::BalsaParser parser_;
 };
 
 } // namespace Http11Connect
